+++ packages/orq-rc/RELEASES.md
@@ -369,3 +369,13 @@ Based on:
 - [typescript v3.1.0-rc.57] packages/orq-rc
 ### Releases
 - [NPM v3.1.0-rc.57] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.57 - packages/orq-rc
+
+## 2025-02-11 09:03:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.488.0 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0-rc.58] packages/orq-rc
+### Releases
+- [NPM v3.1.0-rc.58] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.58 - packages/orq-rc